Adding logout after the user has deleted his profile account.

+++ src/main/java/be/intecbrussel/iddblog/controller/RegisteredVisitorController.java
@@ -25,6 +25,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.io.IOException;
+import java.security.Principal;
 import java.util.Base64;
 import java.util.HashMap;
 import java.util.Map;
@@ -221,9 +222,19 @@ public class RegisteredVisitorController implements HandlerExceptionResolver {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteRegisteredVisitor(@PathVariable("id") long id) {
+    public String deleteRegisteredVisitor(@PathVariable("id") long id, Principal principal, HttpServletRequest request,
+                                          HttpServletResponse response) {
         RegisteredVisitor registeredVisitor = registeredVisitorService.findById(id);
+
+        // The user should only access his own data, otherwise redirect to another page
+        if (!principal.getName().equals(registeredVisitor.getUsername())) {
+            return "redirect:/forbidden-page";
+        }
+
         registeredVisitorService.deleteVisitor(registeredVisitor.getUsername());
+
+        // to logout after the user has deleted his own account
+        logout(request,response);
 
         return "redirect:/index";
     }
@@ -278,6 +289,13 @@ public class RegisteredVisitorController implements HandlerExceptionResolver {
         }
 
         return new ModelAndView("registerform", model);
+    }
+
+    private void logout(HttpServletRequest request, HttpServletResponse response) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth != null) {
+            new SecurityContextLogoutHandler().logout(request, response, auth);
+        }
     }
